+++ src/simplevector.c
@@ -23,9 +23,9 @@ DLLEXPORT jl_svec_t *jl_svec(size_t n, ...)
 jl_svec_t *jl_svec1(void *a)
 {
 #ifdef OVERLAP_SVEC_LEN
-    jl_svec_t *v = (jl_svec_t*)alloc_1w();
+    jl_svec_t *v = (jl_svec_t*)jl_gc_alloc_1w();
 #else
-    jl_svec_t *v = (jl_svec_t*)alloc_2w();
+    jl_svec_t *v = (jl_svec_t*)jl_gc_alloc_2w();
 #endif
     jl_set_typeof(v, jl_simplevector_type);
     jl_svec_set_len_unsafe(v, 1);
@@ -36,9 +36,9 @@ jl_svec_t *jl_svec1(void *a)
 jl_svec_t *jl_svec2(void *a, void *b)
 {
 #ifdef OVERLAP_SVEC_LEN
-    jl_svec_t *v = (jl_svec_t*)alloc_2w();
+    jl_svec_t *v = (jl_svec_t*)jl_gc_alloc_2w();
 #else
-    jl_svec_t *v = (jl_svec_t*)alloc_3w();
+    jl_svec_t *v = (jl_svec_t*)jl_gc_alloc_3w();
 #endif
     jl_set_typeof(v, jl_simplevector_type);
     jl_svec_set_len_unsafe(v, 2);